REFACTOR:Changing the name back to components
