+++ app/events/components/EventCard.tsx
