+++ app/events/components/EventClient.tsx
