+++ app/events/components/SearchBar.tsx
